refactor: fixtures and types
