+++ test/shared/helpers/fixtures/beefyBiswapLpVaultOperatorFixture.ts
@@ -1,0 +1,1 @@
+// exemple new operator file
