+++ test/shared/helpers/fixtures/beefyUniswapLpVaultOperatorFixture.ts
@@ -1,0 +1,1 @@
+// exemple new operator file
