+++ test/shared/helpers/fixtures/zeroExOperatorFixture.ts
@@ -1,0 +1,46 @@
+import { Interface } from "@ethersproject/abi";
+import { Fixture } from "ethereum-waffle";
+import { Wallet } from "ethers";
+import { ethers } from "hardhat";
+import { DummyRouter, MockERC20, TestableOperatorCaller, ZeroExOperator } from "../../../typechain";
+import { appendDecimals } from "../../helpers";
+import { ActorFixture } from "../actors";
+
+
+export type ZeroExOperatorFixture = {
+    zeroExOperator: ZeroExOperator;
+    dummyRouter: DummyRouter;
+    dummyRouterInterface: Interface;
+    mockUNI: MockERC20;
+    mockDAI: MockERC20;
+    testableOperatorCaller: TestableOperatorCaller;
+};
+
+export const zeroExOperatorFixture: Fixture<ZeroExOperatorFixture> = async (wallets, provider) => {
+    const signer = new ActorFixture(wallets as Wallet[], provider).zeroExOperatorOwner();
+
+    const dummyRouterFactory = await ethers.getContractFactory("DummyRouter");
+    const dummyRouter = await dummyRouterFactory.connect(signer).deploy();
+
+    const dummyRouterInterface = dummyRouter.interface;
+
+    const zeroExOperatorFactory = await ethers.getContractFactory("ZeroExOperator");
+    const zeroExOperator = await zeroExOperatorFactory.connect(signer).deploy(dummyRouter.address);
+
+    const mockERC20Factory = await ethers.getContractFactory("MockERC20");
+    const mockUNI = await mockERC20Factory.deploy("Mocked UNI", "UNI", appendDecimals(3000000));
+    await mockUNI.deployed();
+    const mockDAI = await mockERC20Factory.deploy("Mocked DAI", "DAI", appendDecimals(3000000));
+    await mockDAI.deployed();
+
+    await mockUNI.transfer(dummyRouter.address, appendDecimals(1000));
+    await mockDAI.transfer(dummyRouter.address, appendDecimals(1000));
+
+    const testableOperatorCallerFactory = await ethers.getContractFactory("TestableOperatorCaller");
+    const testableOperatorCaller = await testableOperatorCallerFactory.connect(signer).deploy(zeroExOperator.address);
+
+    await mockUNI.transfer(testableOperatorCaller.address, appendDecimals(1000));
+    await mockDAI.transfer(testableOperatorCaller.address, appendDecimals(1000));
+
+    return { zeroExOperator, dummyRouter, dummyRouterInterface, mockUNI, mockDAI, testableOperatorCaller };
+};
