+++ test/shared/types/OperatorResolverFixture.ts
@@ -1,0 +1,3 @@
+import { OperatorResolver } from "../../../../../typechain";
+
+export type OperatorResolverFixture = { operatorResolver: OperatorResolver };
